+++ test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @platforma-open/milaboratories.table.test
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated settings button icon
+
 ## 0.5.2
 
 ### Patch Changes